+++ src/ProcessInvoke.ts
@@ -1,5 +1,9 @@
 import * as ChildProcess from 'child_process';
 import which = require('which');
+import * as readline from 'readline';
+import ProgressBar = require('progress');
+import { https } from 'follow-redirects';
+import * as fse from 'fs-extra';
 import { UserSettingsPath } from './user-settings/UserSettingsPath';
 
 /**
@@ -57,15 +61,93 @@ export async function restorePackages(packageManager: string): Promise<void> {
     }
 }
 
+const isWindows = process.platform === 'win32';
+
+export async function askUserDownloadMkcert(): Promise<boolean> {
+    const rl = readline.createInterface({
+        input: process.stdin,
+        output: process.stdout
+    });
+
+    return await new Promise<boolean>((ok) => {
+        rl.question('mkcert is not found in CLI path. Download mkcert? [Y/n] ', answer => {
+            answer = answer.trim().toUpperCase()[0];
+            if (!answer || answer === 'Y') {
+                ok(true);
+            } else {
+                ok(false);
+            }
+            rl.close();
+        });
+    });
+}
+
+export async function downloadMkcert(): Promise<void> {
+    const downloadUri = 'https://github.com/FiloSottile/mkcert/releases/download/v1.4.1/mkcert-v1.4.1-windows-amd64.exe';
+
+    const binary = await new Promise<Buffer>((ok, reject) => {
+        const data: Uint8Array[] = [];
+
+        https.get(downloadUri, res => {
+            if (res.statusCode !== 200) {
+                reject(new Error('Error downloading mkcert, status code: ' + res.statusCode));
+                return;
+            }
+
+            const contentLength = res.headers['content-length'];
+            let bar: ProgressBar | undefined;
+
+            if (contentLength) {
+                bar = new ProgressBar('Downloading mkcert [:bar] :percent ETA: :etas', {
+                    complete: '=',
+                    incomplete: ' ',
+                    width: 50,
+                    total: parseInt(contentLength)
+                });
+            }
+
+            res.on('data', chunk => {
+                data.push(chunk);
+                bar && bar.tick(chunk.length);
+            }).on('end', () => {
+                ok(Buffer.concat(data));
+            }).on('error', err => {
+                reject(err);
+                bar && bar.terminate();
+            });
+        }).end();
+    });
+
+    await fse.writeFile(UserSettingsPath.mkcertExe, binary)
+    console.log(`\nSuccessfully downloaded mkcert to: ${UserSettingsPath.mkcertExe}\n`);
+}
+
 export async function setupHttps(): Promise<void> {
     const mkcertExists = await isCommandExist('mkcert');
-    if (!mkcertExists) {
+    let mkcertCommand = 'mkcert';
+    let userMkcertExists = false;
+
+    if (!mkcertExists && isWindows) {
+        userMkcertExists = await fse.pathExists(UserSettingsPath.mkcertExe)
+        if (userMkcertExists === false) {
+            const consent = await askUserDownloadMkcert();
+            if (consent === true) {
+                await downloadMkcert();
+                userMkcertExists = true;
+            }
+        }
+        if (userMkcertExists) {
+            mkcertCommand = UserSettingsPath.mkcertExe;
+        }
+    }
+
+    if (!mkcertExists && !userMkcertExists) {
         throw new Error('mkcert is not found in CLI path.\nPlease install mkcert to setup HTTPS: https://github.com/FiloSottile/mkcert');
     }
 
-    execWithConsoleOutput('mkcert -install');
+    const installCommand = mkcertCommand + ' -install';
+    execWithConsoleOutput(installCommand);
 
-    let mkcertCommand = "mkcert";
     mkcertCommand += ` -key-file ${UserSettingsPath.keyFile}`;
     mkcertCommand += ` -cert-file ${UserSettingsPath.certFile}`;
     mkcertCommand += ` "*.test" localhost 127.0.0.1 ::1`;
